# docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {},
    "license": {}
  },
  "paths": {
    "/api/v1/tags": {
      "post": {
        "produces": [
          "application/json"
        ],
        "summary": "新增文章标签",
        "parameters": [
          {
            "type": "string",
            "description": "Name",
            "name": "name",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "State",
            "name": "state",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "CreatedBy",
            "name": "created_by",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/tags/{id}": {
      "put": {
        "produces": [
          "application/json"
        ],
        "summary": "修改文章标签",
        "parameters": [
          {
            "type": "integer",
            "description": "ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "ID",
            "name": "name",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "State",
            "name": "state",
            "in": "query"
          },
          {
            "type": "string",
            "description": "ModifiedBy",
            "name": "modified_by",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "models.Tag": {
      "type": "object",
      "properties": {
        "created_by": {
          "type": "string"
        },
        "created_on": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "modified_by": {
          "type": "string"
        },
        "modified_on": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "state": {
          "type": "integer"
        }
      }
    }
  }
}